+++ librz/asm/arch/sh/regs.h
@@ -1,13 +1,13 @@
 // SPDX-FileCopyrightText: 2022 Dhruv Maroo <dhruvmaru007@gmail.com>
 // SPDX-License-Identifier: LGPL-3.0-only
 
-#ifndef RZ_ASM_SH_DISASSEMBLER_INC
-#define RZ_ASM_SH_DISASSEMBLER_INC
+#ifndef RZ_SH_REGS_H
+#define RZ_SH_REGS_H
 
 /**
  * All registers
  */
-const char *sh_registers[] = {
+static const char *sh_registers[] = {
 	"r0", "r1", "r2", "r3", "r4", "r5", "r6", "r7",
 	"r8", "r9", "r10", "r11", "r12", "r13", "r14", "r15", "pc",
 	"sr", "gbr", "ssr", "spc", "sgr", "dbr", "vbr", "mach", "macl",
@@ -18,4 +18,4 @@ const char *sh_registers[] = {
 	"xf8", "xf9", "xf10", "xf11", "xf12", "xf13", "xf14", "xf15"
 };
 
-#endif // RZ_ASM_SH_DISASSEMBLER_INC
+#endif // RZ_SH_REGS_H
