+++ librz/parse/p/parse_sh_pseudo.c
@@ -22,13 +22,13 @@ static const RzPseudoGrammar sh_lexicon[] = {
 	RZ_PSEUDO_DEFINE_GRAMMAR("and", "2 &= 1"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("and.b", "2 &= 1"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("bf", "if (!t) goto 1"),
-	RZ_PSEUDO_DEFINE_GRAMMAR("bf.s", "if (!t) goto 1"),
+	RZ_PSEUDO_DEFINE_GRAMMAR("bf/s", "if (!t) goto 1"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("bra", "goto 1"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("brk", "_break_exception ()"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("bsr", "1 ()"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("bsrf", "1 ()"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("bt", "if (t) goto 1"),
-	RZ_PSEUDO_DEFINE_GRAMMAR("bt.s", "if (t) goto 1"),
+	RZ_PSEUDO_DEFINE_GRAMMAR("bt/s", "if (t) goto 1"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("clrmac", "_clrmac ()"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("clrs", "_clrs ()"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("clrt", "_clrt ()"),
@@ -99,11 +99,11 @@ static const RzPseudoGrammar sh_lexicon[] = {
 	RZ_PSEUDO_DEFINE_GRAMMAR("shar", "1 >>= #1"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shld", "2 = 1 >= 0 ? 2 << 1 | 2 >> (#31 - 1)"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shll", "1 <<= #1"),
-	RZ_PSEUDO_DEFINE_GRAMMAR("shll#16", "1 <<= #16"),
+	RZ_PSEUDO_DEFINE_GRAMMAR("shll16", "1 <<= #16"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shll2", "1 <<= #2"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shll8", "1 <<= #8"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shlr", "1 >>= #1"),
-	RZ_PSEUDO_DEFINE_GRAMMAR("shlr#16", "1 >>= #16"),
+	RZ_PSEUDO_DEFINE_GRAMMAR("shlr16", "1 >>= #16"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shlr2", "1 >>= #2"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("shlr8", "1 >>= #8"),
 	RZ_PSEUDO_DEFINE_GRAMMAR("sleep", "_halt ()"),
